editor: move the profiler console commands out of the web app
